+++ packages/react-aria-components/src/RSPContexts.ts
@@ -10,6 +10,8 @@
  * governing permissions and limitations under the License.
  */
 
+import {AriaLabelingProps, DOMProps, FocusableElement, FocusEvents, KeyboardEvents, Node, ValueBase} from '@react-types/shared';
+import {AriaTextFieldProps} from '@react-aria/textfield';
 import {CheckboxProps} from './Checkbox';
 import {ColorAreaProps} from './ColorArea';
 import {ColorFieldProps} from './ColorField';
@@ -31,3 +33,20 @@ export const ColorFieldContext = createContext<ContextValue<ColorFieldProps, HTM
 export const ColorSliderContext = createContext<ContextValue<Partial<ColorSliderProps>, HTMLDivElement>>(null);
 export const ColorWheelContext = createContext<ContextValue<Partial<ColorWheelProps>, HTMLDivElement>>(null);
 export const HeadingContext = createContext<ContextValue<HeadingProps, HTMLHeadingElement>>({});
+
+export interface SelectableCollectionContextValue<T> extends DOMProps, AriaLabelingProps {
+  filter?: (nodeTextValue: string, node: Node<T>) => boolean,
+  /** Whether the collection items should use virtual focus instead of being focused directly. */
+  shouldUseVirtualFocus?: boolean,
+  /** Whether typeahead is disabled. */
+  disallowTypeAhead?: boolean
+}
+interface FieldInputContextValue<T = FocusableElement> extends
+  DOMProps,
+  FocusEvents<T>,
+  KeyboardEvents,
+  Pick<ValueBase<string>, 'onChange' | 'value'>,
+  Pick<AriaTextFieldProps, 'enterKeyHint' | 'aria-controls' | 'aria-autocomplete' | 'aria-activedescendant' | 'spellCheck' | 'autoCorrect' | 'autoComplete'> {}
+
+export const SelectableCollectionContext = createContext<ContextValue<SelectableCollectionContextValue<any>, HTMLElement>>(null);
+export const FieldInputContext = createContext<ContextValue<FieldInputContextValue, FocusableElement>>(null);
